File Structure Refactord
