+++ .htaccess
@@ -1,0 +1,1 @@
+ErrorDocument 404 @/pages/404.tsx
