+++ src/components/atoms/animated-shiny-text.tsx
