+++ src/components/atoms/icon-cloud.tsx
